+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.6.0
+
+👀
+
+## 0.5.3
+
+- Added docstring to `SQLAlchemy()`.
+- More type fixes.
+
 ## 0.5.2
 
 - Fixed poorly handled merge conflict leaving `.sqlalchemy` modulem unusable